requirements.txt
azure-iot-device
dotenv